+++ src/pages/bookshelf/BookshelfItemPage.vue
@@ -46,11 +46,17 @@
 
           <book-details :book="book" />
 
+          <bookshelf-rarimo-message
+            v-if="!isValidChain && provider.chainId"
+            :source-chain="provider.chainId"
+            :target-chain="$config.DEFAULT_CHAIN_ID"
+          />
+
           <app-button
             v-if="provider.isConnected"
             class="bookshelf-item-page__purchase-btn"
             size="small"
-            :text="$t('bookshelf-item-page.purchase-btn')"
+            :text="buyButtonText"
             @click="isPurchaseModalShown = true"
           />
 
@@ -67,6 +73,7 @@
           v-if="book"
           v-model:is-shown="isPurchaseModalShown"
           :book="book"
+          :is-valid-chain="isValidChain"
           @submit="submit"
         />
 
@@ -96,15 +103,15 @@ import {
   Icon,
 } from '@/common'
 
-import { BookDetails } from '@/pages/bookshelf'
-import { ref, watch, computed } from 'vue'
+import { BookDetails, BookshelfRarimoMessage } from '@/pages/bookshelf'
+import { ref, computed } from 'vue'
 
 import { ErrorHandler, getBlockExplorerLink } from '@/helpers'
 
 import { useBooks } from '@/composables'
 import { FullBookInfo } from '@/types'
 
-import { useWeb3ProvidersStore, useNetworksStore } from '@/store'
+import { useWeb3ProvidersStore } from '@/store'
 import { router } from '@/router'
 import { ROUTE_NAMES } from '@/enums'
 import { useI18n } from 'vue-i18n'
@@ -115,7 +122,6 @@ const props = defineProps<{
 
 const { t } = useI18n()
 
-const networkStore = useNetworksStore()
 const web3Store = useWeb3ProvidersStore()
 const provider = computed(() => web3Store.provider)
 
@@ -128,6 +134,21 @@ const linkToTx = ref('')
 const { getBookById } = useBooks()
 
 const book = ref<FullBookInfo | undefined>()
+const isValidChain = computed(() =>
+  Boolean(
+    book.value &&
+      book.value.networks.some(
+        network =>
+          network.attributes.chain_id === Number(provider.value.chainId),
+      ),
+  ),
+)
+
+const buyButtonText = computed(() =>
+  isValidChain.value
+    ? t('bookshelf-item-page.purchase-btn')
+    : t('bookshelf-item-page.rarimo-btn'),
+)
 
 const submit = async (message?: string) => {
   try {
@@ -153,7 +174,7 @@ const submit = async (message?: string) => {
 const init = async () => {
   isLoaded.value = false
   try {
-    const data = await getBookById(props.id, provider.value.chainId)
+    const data = await getBookById(props.id, provider.value.chainId!)
 
     book.value = data
   } catch (error) {
@@ -162,29 +183,6 @@ const init = async () => {
   }
   isLoaded.value = true
 }
-
-watch(
-  () => provider.value.isConnected,
-  value => {
-    if (!value) {
-      isPurchaseModalShown.value = false
-    }
-  },
-)
-
-watch(
-  () => provider.value.chainId,
-  () => {
-    if (
-      !networkStore.list.some(
-        network => network.chain_id === Number(provider.value.chainId),
-      )
-    )
-      return
-
-    init()
-  },
-)
 
 init()
 </script>
@@ -257,7 +255,6 @@ init()
 }
 
 .bookshelf-item-page__purchase-btn {
-  text-transform: uppercase;
   width: 100%;
   height: toRem(60);
   font-size: toRem(24);