+++ src/pages/my-nfts/MyNFTsPage.vue
@@ -4,8 +4,11 @@
       {{ $t('my-nfts-page.title') }}
     </h3>
     <loader v-if="isLoading" />
-    <nft-list v-else-if="provider.isConnected" :total-amount="totalAmount" />
-    <my-nfts-no-data v-else is-not-connected />
+    <template v-else>
+      <my-nfts-no-data v-if="noDataScheme" :scheme="noDataScheme" />
+      <nft-list v-else-if="totalAmount" :total-amount="totalAmount" />
+    </template>
+
     <img class="my-nfts-page__background" src="/images/fancy-lines.png" />
   </div>
 </template>
@@ -14,12 +17,26 @@
 import { Loader } from '@/common'
 import { MyNftsNoData, NftList } from '@/pages/my-nfts'
 import { computed, ref, watch } from 'vue'
-import { useWeb3ProvidersStore } from '@/store'
+import { useWeb3ProvidersStore, useNetworksStore } from '@/store'
 import { useBooks } from '@/composables'
 import { ErrorHandler } from '@/helpers'
 
+const networkStore = useNetworksStore()
 const web3ProvidersStore = useWeb3ProvidersStore()
 const provider = computed(() => web3ProvidersStore.provider)
+
+const isValidChain = computed(() =>
+  networkStore.list.some(
+    i => Number(i.chain_id) === Number(provider.value.chainId),
+  ),
+)
+
+const noDataScheme = computed(() => {
+  if (!isValidChain.value) return 'wrong-chain'
+  else if (!provider.value.isConnected) return 'not-connected'
+
+  return ''
+})
 
 const totalAmount = ref(-1)
 const isLoading = ref(false)