+++ src/pages/my-nfts/MyNftsNoData.vue
@@ -5,8 +5,8 @@
       src="/images/books-no-data.png"
       alt="Books no data image"
     />
-    <p v-if="isNotConnected" class="my-nfts-no-data__message">
-      {{ $t('my-nfts-no-data.connect-metamask-message') }}
+    <p v-if="errorMessage" class="my-nfts-no-data__message">
+      {{ errorMessage }}
     </p>
     <template v-else>
       <p class="my-nfts-no-data__message">
@@ -24,6 +24,13 @@
       @click="provider.connect"
     />
     <app-button
+      v-else-if="isWrongChain"
+      class="my-nfts-no-data__marketplace-btn"
+      :text="$t('my-nfts-no-data.switch-chain-btn')"
+      :icon-left="$icons.refresh"
+      @click="switchNetwork($config.DEFAULT_CHAIN_ID)"
+    />
+    <app-button
       v-else
       class="my-nfts-no-data__marketplace-btn"
       size="large"
@@ -38,18 +45,33 @@
 import { computed } from 'vue'
 import { AppButton } from '@/common'
 import { useWeb3ProvidersStore } from '@/store'
+import { switchNetwork } from '@/helpers'
+import { useI18n } from 'vue-i18n'
+
+type SCHEME = 'not-connected' | 'wrong-chain' | 'default'
+
+const { t } = useI18n()
 
 const web3ProvidersStore = useWeb3ProvidersStore()
 const provider = computed(() => web3ProvidersStore.provider)
 
-withDefaults(
+const props = withDefaults(
   defineProps<{
-    isNotConnected?: boolean
+    scheme?: SCHEME
   }>(),
   {
-    isNotConnected: false,
+    scheme: 'default',
   },
 )
+
+const isNotConnected = computed(() => props.scheme === 'not-connected')
+const isWrongChain = computed(() => props.scheme === 'wrong-chain')
+const errorMessage = computed(() => {
+  if (isNotConnected.value) return t('my-nfts-no-data.connect-metamask-message')
+  else if (isWrongChain.value) return t('my-nfts-no-data.wrong-chain-message')
+
+  return ''
+})
 </script>
 
 <style lang="scss" scoped>